refactor: 날짜/시간 포맷 순수 함수를 chatDate 유틸로 분리
